Remove hidden global variable and procedure Register
The hidden global variable caused several hooks to be attached to the
wrong plug-in when loading multiple Ada plug-ins. In order to provide
the plug-in pointer to callbacks, the pointer is put inside a record
that is allocated on the heap. This record will be deallocated by
WeeChat when the hook is removed (when the plug-in is unloaded).

Procedure Register has been removed because WeeChat uses the exported
weechat_* symbols to display certain metadata of the plug-in. Even
though the plug-in metadata is part of the plug-in data structure.
Plug-ins must now copy some boilerplate code to export this metadata.

Signed-off-by: onox <denkpadje@gmail.com>